// client-monorepo/libs/applets/insurance/src/lib/app/data-access/enums/motor-third-party-page-title.enum.ts
export enum MotorThirdPartyPageTitleEnum {
  MODEL_BUILD_YEAR = 'Model',
  ExInsurerMotor = 'ExInsurer',
  ExInsurerMotorDate = 'ExInsurerMotorDate',
  ExInsurerMotorInfo = 'ExInsurerMotorInfo',
  PriceCardList = 'PriceCardList',
  GoToCheckout = 'GoToCheckout',
  Checkout = 'Checkout',
  Payment = 'Payment',
  UserInfo = 'UserInfo',
  UserAddress = 'UserAddress',
  UploadDocument = 'UploadDocument',
  GetPolicyMethod = 'GetPolicyMethod',
  CompleteOrder = 'CompleteOrder',
  OrderState = 'OrderState',
  AdditionalUploadDocument = 'AdditionalUploadDocument',
}

export const MOTOR_THIRD_PARTY_TITLES_TRANSLATOR = {
  [MotorThirdPartyPageTitleEnum.MODEL_BUILD_YEAR]: 'مشخصات موتور سیکلت',
  [MotorThirdPartyPageTitleEnum.ExInsurerMotor]: 'مشخصات بیمه‌نامه',
  [MotorThirdPartyPageTitleEnum.ExInsurerMotorDate]: 'مشخصات بیمه‌نامه',
  [MotorThirdPartyPageTitleEnum.ExInsurerMotorInfo]: 'تخفیف و خسارت بیمه‌نامه',
  [MotorThirdPartyPageTitleEnum.PriceCardList]: 'قیمت‌های بیمه موتور سیکلت',
  [MotorThirdPartyPageTitleEnum.GoToCheckout]: 'انتقال برای انتخاب روش پرداخت',
  [MotorThirdPartyPageTitleEnum.Checkout]: 'انتخاب روش پرداخت',
  [MotorThirdPartyPageTitleEnum.Payment]: 'پرداخت',
  [MotorThirdPartyPageTitleEnum.UserInfo]: 'تکمیل اطلاعات',
  [MotorThirdPartyPageTitleEnum.UserAddress]: 'تکمیل اطلاعات',
  [MotorThirdPartyPageTitleEnum.UploadDocument]: 'آپلود مدارک',
  [MotorThirdPartyPageTitleEnum.GetPolicyMethod]: 'شیوه ی دریافت بیمه نامه',
  [MotorThirdPartyPageTitleEnum.CompleteOrder]: 'تکمیل سفارش',
  [MotorThirdPartyPageTitleEnum.OrderState]: 'وضعیت سفارش',
  [MotorThirdPartyPageTitleEnum.AdditionalUploadDocument]: 'آپلود مدارک اضافه',
};